Add St Helena exception to courier_fee_info partial
We've done this by introducing the `#courier_service?` method. We can't
be sure but the lack of a courier service appears to be the thing that differentiates St Helena from all other countries.

This moves the content in the `if` branch of the _cost conditional
toward that in the `else` branch. Once identical we'll be able to remove
the conditional entirely.