Provide a text version of a question page
Previously it was only possible to render a text version of a landing page or an
outcome page.

This will make it possible to generate regression test artefacts for question
pages in text form.

As per the other page types I have not annotated each "attribute" of the
question. And as per the other page types each "attribute" is always rendered
whether or not it is blank. Unfortunately this makes the text version of some
question pages not terribly readable.

However, I think it's more valuable to keep the text version of the three page
types consistent with each other. And this should be good enough to use the new
question page format for regression test artefacts which is what I'm focussed on
at the moment.

Note that I've used bullet points to list any options associated with the
question. This seemed like the simplest way to represent them in text form.