Render regression test artefacts for question pages as text
I haven't updated the artefacts themselves in this commit, because I'm planning
to do that using one commit per flow. Although this means that the regression
tests are all failing after this commit, I think the approach will make it
easier to see what's going on.

In making this change I have double-checked that all possible "attributes" for
all question types are included in the text version of a question page:

* title
* body
* hint
* options (each with value & label)
* label
* suffix_label
* post_body

The point of making this change is to make regression test artefacts for
question pages less brittle. However, it's worth noting that in making this
change we are also losing a bit of test coverage for the "previous answers"
section. The latter functionality is tested elsewhere, although perhaps not in
quite such an end-to-end fashion. As long as we are conscious of this when
making changes in this area, I think this is an acceptable risk.