Disable AJAX on some Smart Answers
This disables the JS which loads questions via AJAX.
We would like to remove the AJAX altogether
as it is not making things much quicker
and it introduces maintability issues and bugs.
To be reassured that this will not worsen the user experience,
we only disable it on some Smart Answers.
The list of Smart Answers was provided by Dan
who chose some which are not visited too often or too seldom.