Update Consular fees table with Civil Union
Here I have added to the marriage abroad consular fees table options another option
(receiving_notice_of_civil_union). At the moment this option is exactly the same as
receiving_notice_of_marriage option, they also share the same price.

The rational for adding this option is to allow for more accurate content to be used in the fee table.
As explained by the Content designer that this is a very important and taken seriously in Italy.

In the services YAML file I have added same sex option for ceremony and third country as specified in the document.